requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
click-default-group==1.2.4
colorama==0.4.6
fastapi==0.116.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pluggy==1.6.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
sniffio==1.3.1
sqlite-fts4==1.0.3
sqlite-utils==3.38
starlette==0.48.0
tabulate==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
